[SMAGENT-1082] UI reports a wrong memory.bytes.used metric for sdjagent processes
During startup, a process's thread creation event may be missed (if it
happens in between /proc scan and scap_start_capture). In some cases,
this can result in misbehaving, namely, in mistaking the thread as a
process. To overcome the issue, during a short interval after startup,
if agent gets an event for a non-scanned thread, it performs a second
scan